Fix Verus compilation, update gate+README, expand Zephyr CI
Verus verification:
  - Fix 30 compilation errors in 6 agent-written modules
    (atomic: old() shadowing, heap/kheap/stack_config: bitwise-on-int,
     userspace: type inference, mem_domain: spec return type)
  - VERUS_SRCS changed to glob (never stale again)
  - poll.rs excluded (mutable array indexing not yet supported by Verus)
  - sched.rs excluded (function pointer types not yet supported)
  - Remaining issue: #[derive(Clone)] inside verus!{} blocks generates
    exec code Verus can't handle. Original 6 modules (sem/mutex/condvar/
    msgq/stack/pipe) verify clean. Newer modules need derives moved
    outside verus!{} blocks.

HONEST STATUS:
  - Modules that pass Verus: sem, mutex, condvar, msgq, stack, pipe,
    wait_queue, thread, priority, error (the original core)
  - Modules with Verus syntax issues: 28 newer agent-written modules
    (derives inside verus!{}, bitwise ops, function pointers)
  - All 38 modules pass cargo test (995 tests)
  - All 38 modules have ASPICE traceability (660 artifacts)

Gate test: updated to check all 38 modules
README: rewritten with full module table + known gaps section
Zephyr CI: expanded from 13 to 20 test suites
Multi-arch: M3+M4+M33 all build

+++ src/mem_domain.rs
@@ -87,7 +87,7 @@ impl MemPartition {
     }
 
     /// End address (exclusive) — spec only, avoids overflow in exec code.
-    pub open spec fn end_spec(&self) -> u64 {
+    pub open spec fn end_spec(&self) -> int {
         self.start as u64 + self.size as u64
     }
